Move logic to View widget
