+++ packages/flutter/lib/src/widgets/view.dart
@@ -193,24 +193,43 @@ class _ViewState extends State<View> with WidgetsBindingObserver {
     debugLabel: kReleaseMode ? null : 'View Scope',
   );
   final FocusTraversalPolicy _policy = ReadingOrderTraversalPolicy();
+  bool _viewHasFocus = false;
 
   @override
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
+    _scopeNode.addListener(_scopeFocusChangeListener);
   }
 
   @override
   void dispose() {
     WidgetsBinding.instance.removeObserver(this);
+    _scopeNode.removeListener(_scopeFocusChangeListener);
     _scopeNode.dispose();
     super.dispose();
   }
 
+  void _scopeFocusChangeListener() {
+    if (_viewHasFocus == _scopeNode.hasFocus) {
+      return;
+    }
+    // Scope focus changed, and it doesn't match the view focus, so inform the
+    // view so it knows to change its focus.
+    WidgetsBinding.instance.platformDispatcher.requestViewFocusChange(
+      direction: ViewFocusDirection.forward,
+      state: _scopeNode.hasFocus ? ViewFocusState.focused : ViewFocusState.unfocused,
+      viewId: widget.view.viewId,
+    );
+  }
+
   @override
   void didChangeViewFocus(ViewFocusEvent event) {
+    _viewHasFocus = switch (event.state) {
+      ViewFocusState.focused => event.viewId == widget.view.viewId,
+      ViewFocusState.unfocused => false,
+    };
     if (event.viewId != widget.view.viewId) {
-      // The event is not pertinent to this view.
       return;
     }
     FocusNode nextFocus;
